refactor: projects-showcase
