+++ src/page/main-page/ui/project-showcase/lib/constants.tsx
@@ -1,9 +1,4 @@
-export type TCard = {
-  img: string
-  alt: string
-  title: string
-  description: string
-}
+import { TCard } from '@/shared/types/projects'
 
 export const cards: TCard[] = [
   {
