+++ src/page/main-page/ui/project-showcase/style.module.scss
@@ -1,17 +1,5 @@
 @import '@/mixins';
 
-.showcase {
-  padding-top: responsive(60px, 120px);
-  padding-bottom: responsive(60px, 125px);
-  width: 1260px;
-
-  @media (max-width: 1300px) {
-    width: 100%;
-  }
-
-  .title {
-    margin-bottom: responsive(50px, 60px);
-    max-width: 705px;
-    text-align: left;
-  }
+.wrap {
+  @include paddingY(30px, 120px);
 }
