+++ src/shared/types/projects.ts
@@ -1,0 +1,6 @@
+export type TCard = {
+  img: string
+  alt: string
+  title: string
+  description: string
+}
